Solved online and offline modes/ Added SPIFFS
Initially, the device starts in Station mode and tries to connect to the given SSID. This data of SSID and password is persistent in memory. If the device fails to connect to the network, the Wifi Manager portal opens up and is ready to be configured. Once the portal starts then we may give the credentials to connect to the Internet. While in online mode, the device sends the data to the server. In offline mode, the data remains in the chip in SPIFFS file system. Little FS can also be used. I am not much aware of it.