cleaned code and logging is perfect now
Data is getting stored to the chip in a text file, we can have 3MB of space. Removed unused counter variable and rest of comments. Played with multidimensional array to store data into file. Tried SerialJson but it did not work out well, so made a simple file and opened in addition mode.